Clean up season stats helper test files
